feat: migrate Zod to pure TS w/ JSDoc, improve type safety and simplify parsing (#217)
# Description

<!-- Include a summary of the change made and also list the dependencies
that are required if any -->

This PR migrates the repo off of Zod schemas toward pure TS with JSDoc
annotations that can eventually be used to extract metadata for display
in docs.

Generally the external behavior should be identical, with obvious
exceptions where schema entry points are no longer available etc., so
this would constitute a breaking change for consumers relying on those
entrypoints or whose inputs may no longer be valid with new type safety
around video/image options.

I've also updated all the tests to `.ts` so we can be sure the types for
the API are working as we expect going forward.

Although all the tests are passing, the doc generation issue remains
unsolved, causing the repo-wide build to fail, so that will need to be
updated to use a JSDoc parsing tool. However, it should be quite
straightforward to wire that up in place of the previous Zod-embedded
metadata.

For now, I'm opening this against the `beta` branch so the required
JSDoc parsing logic can be added and other changes can be experimented
with before broader consumption.

## Issue Ticket Number

N/A

<!-- Specify above which issue this fixes by referencing the issue
number (`#<ISSUE_NUMBER>`) or issue URL. -->
<!-- Example: Fixes
https://github.com/colbyfayock/cloudinary-util/issues/<ISSUE_NUMBER> -->

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [X] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Fix or improve the documentation
- [X] This change requires a documentation update

# Checklist

(Still a draft, need to figure out how to work around missing schemas in
build)

<!-- These must all be followed and checked. -->

- [ ] I have followed the contributing guidelines of this project as
mentioned in [CONTRIBUTING.md](/CONTRIBUTING.md)
- [ ] I have created an
[issue](https://github.com/colbyfayock/cloudinary-util/issues) ticket
for this PR
- [ ] I have checked to ensure there aren't other open [Pull
Requests](https://github.com/colbyfayock/cloudinary-util/pulls) for the
same update/change?
- [ ] I have performed a self-review of my own code
- [ ] I have run tests locally to ensure they all pass
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes needed to the documentation

BREAKING CHANGES: removes zod schemas, refactors type system